Agentic_AI/AI/AI_Projects/YouTube-Blog-Generator-main/requirement.txt
langchain
python-dotenv
Langchain-openai
langchain-core
langchain_community
bs4
faiss-cpu
PyPDF2
streamlit
Langchain_groq
Langchain_huggingface
chromadb
pypdf
arxiv
pymupdf
wikipedia
lxml
langgraph
langgraph-cli[inmem]
youtube_transcript_api
flask
pytube
streamlit_extras.add_vertical_space
streamlit_extras.mention